pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel", "setuptools-scm>=8"]
build-backend = "setuptools.build_meta"

[project]
name = "project-copilot-plusng"
dynamic = ["version"]
description = "Project Copilot Starter Kit – Plus NG"
requires-python = ">=3.12"
dependencies = [
    "pytest>=8.0.0",
    "requests>=2.32.0",
    "streamlit>=1.32.0",
    "PyYAML>=6.0.1",
    "pypdf>=4.0.0",
    "python-docx>=1.1.0"
]

[tool.setuptools]
package-dir = {"" = "."}

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

[tool.setuptools_scm]
version_scheme = "no-guess-dev"
local_scheme = "no-local-version"
tag_regex = "^v(?P<version>\\d+\\.\\d+\\.\\d+)$"


